BoundingBox options, Entity velocity, resizable option
Added an option to make the window non-resizable or resizable. Changed
the preferred movement system for entities to the velocity system,
favoring the future implementation of AI entities and movement based
around reversing movement for collisions. Last of all made  a change to
the BoundingBox class to make entities non-collidable so they can be
drawn over without disabling mechanics such as picking up items.